AppInfo refactoring and preparations
