+++ lib/app/common/app_page/app_loading_page.dart
@@ -80,7 +80,7 @@ class AppLoadingPage extends StatelessWidget {
       baseColor: shimmerBase,
       highlightColor: shimmerHighLight,
       child: const BorderContainer(
-        child: SizedBox(height: 150),
+        child: SizedBox(height: 173),
       ),
     );
