+++ lib/l10n/app_en.arb
@@ -168,5 +168,6 @@
         }
     }
   },
-  "gallery": "Gallery"
+  "gallery": "Gallery",
+  "additionalInformation": "Additional Information"
 }
